\name{LargeScaleLinReg-package}
\alias{LargeScaleLinReg-package}
\alias{LargeScaleLinReg}
\docType{package}
\title{
\packageTitle{LargeScaleLinReg}
}
\description{
\packageDescription{LargeScaleLinReg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LargeScaleLinReg}
\packageIndices{LargeScaleLinReg}
Routine to do multiple linear regressions, XB = Y,
where a large number of columns of the X matrix are constant
and a small number of columns in X change and/or Y changes.
}
\author{
\packageAuthor{LargeScaleLinReg}

Maintainer: \packageMaintainer{LargeScaleLinReg}
}
\references{
In work
}
\keyword{ package }
\seealso{
}
\examples{
LSLinReg(xl, xr, y)
}
