% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSLinReg.R
\name{LSLinReg}
\alias{LSLinReg}
\title{Function to do large scale linear regression}
\usage{
LSLinReg(xl, xr, xrcol = 1L, y)
}
\arguments{
\item{xl}{A matrix with the columns of x that do not change. If the
intercept term is needed, there must be a columns of 1's
in the matrix.}

\item{xr}{A matrix with the values for the columns that have values
changing. All possible values for xr need to be concatenated
into one matrix.}

\item{xrcol}{The number of columns in each xr matrix. The number of columns
in the xr matrix passed to the function must be a multiple
of xrcol}

\item{y}{A matrix with all the possible values of y. Each column
contains one possible value for y.}
}
\value{
A matrix with the beta values for each model. The first n
columns will be for the first value of xr where n is the
number of columns in x.
}
\description{
Function to do multiple linear regressions, XB = Y,
where X has a large number of columns that do not change
and a small number of columns that change and/or Y changes.
}
